django-todo-list/api/requirements.txt
appdirs==1.4.3
asgiref==3.2.5
certifi==2019.11.28
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.5.0
pipenv==2018.11.26
protobuf==3.10.0
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.10
virtualenv-clone==0.5.3
whitenoise==5.0.1
zipp==3.1.0
